chore: prepare mcp npm release
