+++ packages/luau-bundler/src/cli.ts
@@ -1,4 +1,4 @@
-#!/usr/bin/env bun
+#!/usr/bin/env node
 
 /**
  * CLI entry point for the Luau bundler.
