+++ packages/mcp/README.md
@@ -2,12 +2,12 @@
 
 <img src="https://raw.githubusercontent.com/0neShot101/rbxdev-ls/main/packages/vscode/icon.png" alt="rbxdev-ls logo" width="160">
 
-<h1>@0neshot101/rbxdev-mcp</h1>
+<h1>@oneshot101/rbxdev-mcp</h1>
 
 <p>MCP server that connects AI assistants to live Roblox game instances over a local WebSocket bridge</p>
 
 <p>
-  <a href="https://github.com/0neShot101/rbxdev-ls/packages"><img src="https://img.shields.io/badge/GitHub%20Packages-0.3.6-blue?style=flat-square" alt="GitHub Packages"></a>
+  <a href="https://www.npmjs.com/package/@oneshot101/rbxdev-mcp"><img src="https://img.shields.io/npm/v/@oneshot101/rbxdev-mcp?style=flat-square" alt="npm"></a>
   <img src="https://img.shields.io/badge/node-%3E%3D18-brightgreen?style=flat-square" alt="Node.js >= 18">
   <img src="https://img.shields.io/badge/protocol-MCP-blue?style=flat-square" alt="MCP">
   <a href="https://github.com/0neShot101/rbxdev-ls/blob/main/LICENSE"><img src="https://img.shields.io/github/license/0neShot101/rbxdev-ls?style=flat-square" alt="MIT License"></a>
@@ -19,7 +19,7 @@
 
 ## Overview
 
-`@0neshot101/rbxdev-mcp` is a [Model Context Protocol](https://modelcontextprotocol.io) server that gives AI assistants &mdash; Claude, Cursor, Windsurf, and any other MCP-compatible tool &mdash; direct access to a running Roblox game. It exposes 16 tools and 3 resources that let an AI browse the game hierarchy, read and write instance properties, execute Luau code, inspect scripts, monitor remote calls, and manipulate the scene graph, all in real time.
+`@oneshot101/rbxdev-mcp` is a [Model Context Protocol](https://modelcontextprotocol.io) server that gives AI assistants &mdash; Claude, Cursor, Windsurf, and any other MCP-compatible tool &mdash; direct access to a running Roblox game. It exposes 19 tools and 3 resources that let an AI browse the game hierarchy, read and write instance properties, execute Luau code, inspect scripts, monitor remote calls, and manipulate the scene graph, all in real time.
 
 The server is part of the [rbxdev-ls](https://github.com/0neShot101/rbxdev-ls) monorepo and depends on `rbxdev-server` (the shared language server core) for its bridge layer. It communicates with a Roblox executor over a local WebSocket connection and speaks MCP over stdio to the AI client. If the rbxdev-ls VS Code extension is already running, the MCP server detects the occupied port and connects as a proxy through the extension's bridge &mdash; both tools share the same executor connection with zero conflicts.
 
@@ -54,7 +54,7 @@ Add this block to your MCP configuration file and restart your AI tool:
   "mcpServers": {
     "rbxdev-roblox": {
       "command": "npx",
-      "args": ["-y", "@0neshot101/rbxdev-mcp"]
+      "args": ["-y", "@oneshot101/rbxdev-mcp"]
     }
   }
 }
@@ -83,7 +83,7 @@ The server buffers up to 1000 console log entries (print, warn, error) locally s
 
 ```mermaid
 graph LR
-    A["AI Assistant<br/>(Claude, Cursor, Windsurf)"] -- stdio --> B["MCP Server<br/>@0neshot101/rbxdev-mcp"]
+    A["AI Assistant<br/>(Claude, Cursor, Windsurf)"] -- stdio --> B["MCP Server<br/>@oneshot101/rbxdev-mcp"]
     B -- WebSocket --> C["Executor Bridge"]
     C -- Roblox API --> D["Roblox Game"]
     E["rbxdev-ls<br/>VS Code Extension"] -. proxy .-> B
@@ -101,22 +101,18 @@ When the VS Code extension is running, the MCP server connects as a proxy client
 
 ### Registry setup
 
-The package is hosted on GitHub Packages. Configure npm to use the GitHub registry for the `@0neshot101` scope by adding this to your `~/.npmrc`:
-
-```
-@0neshot101:registry=https://npm.pkg.github.com
-```
+The package is published to the public npm registry under the `@oneshot101` scope.
 
 ### npx (recommended)
 
-No installation required. The `npx -y @0neshot101/rbxdev-mcp` command in your MCP config downloads and runs the latest version automatically each time the AI tool starts.
+No installation required. The `npx -y @oneshot101/rbxdev-mcp` command in your MCP config downloads and runs the latest version automatically each time the AI tool starts.
 
 ### Global install
 
 If you prefer a persistent installation:
 
 ```bash
-npm install -g @0neshot101/rbxdev-mcp
+npm install -g @oneshot101/rbxdev-mcp
 ```
 
 Then reference the binary directly in your MCP config:
@@ -144,7 +140,7 @@ To use a custom port, add an `env` block to your MCP config:
   "mcpServers": {
     "rbxdev-roblox": {
       "command": "npx",
-      "args": ["-y", "@0neshot101/rbxdev-mcp"],
+      "args": ["-y", "@oneshot101/rbxdev-mcp"],
       "env": {
         "RBXDEV_BRIDGE_PORT": "21325"
       }
@@ -157,7 +153,7 @@ You can also pass `--port 21325` as a CLI argument instead of the environment va
 
 ## Tools
 
-The server registers 16 tools that the AI assistant can call:
+The server registers 19 tools that the AI assistant can call:
 
 <details>
 <summary><strong>Full tool reference</strong></summary>
@@ -180,6 +176,9 @@ The server registers 16 tools that the AI assistant can call:
 | `refresh_game_tree`      | Request a fresh snapshot of the game tree from the executor                                                                |
 | `get_remote_calls`       | View captured RemoteEvent/RemoteFunction calls with reproducible Luau code                                                 |
 | `set_remote_spy_enabled` | Toggle the Remote Spy on or off to start or stop capturing remote calls                                                    |
+| `set_remote_spy_block_list` | Set remotes to block from firing to the server                                                                         |
+| `set_script_source`      | Update script source when connected through Roblox Studio                                                                 |
+| `save_instance`          | Save the full DataModel or a specific instance to a file                                                                  |
 
 </details>
