+++ packages/mcp/src/handlers/tools/shared.ts
@@ -1,0 +1,9 @@
+import type { ExecutorBridge, LogEntry, ToolResult } from 'rbxdev-server';
+
+export interface ToolHandlerContext {
+  bridge: ExecutorBridge;
+  logBuffer: ReadonlyArray<LogEntry>;
+}
+
+export type ToolHandler = (args: unknown, context: ToolHandlerContext) => Promise<ToolResult>;
+export type ToolHandlerMap = Record<string, ToolHandler>;
